improve: パスワードのinput機構をコンポーネント化
